Graph prototype
