+++ package.json
@@ -11,6 +11,7 @@
   },
   "dependencies": {
     "@observablehq/plot": "^0.6.17",
+    "@plasmohq/messaging": "^0.7.2",
     "plasmo": "0.90.5",
     "react": "18.2.0",
     "react-dom": "18.2.0"
