+++ src/contents/inject/nike/graph.ts
@@ -1,21 +1,9 @@
 import * as Plot from "@observablehq/plot"
-import type { StoreId } from "~lib/stores"
-import { getRecordByKey, type StoreDef } from "~lib/useDB"
+import { sendToBackground } from "@plasmohq/messaging"
 
 export type HTMLSVGElement = HTMLElement & SVGElement
 
-// Should remove this, and not require this argument in the future
-// Also this only works for Nike I guess
-const STORE_NAME: StoreId = "nike"
-const STORE_INFO: StoreDef = {
-    keyPath: "key",
-    indexes: [
-        { name: "groupKey", keyPath: "groupKey", unique: false },
-        { name: "productCode", keyPath: "productCode", unique: true }
-    ]
-}
-
 export async function getGraphForItem(key: string) {
-    const priceData = (await getRecordByKey(STORE_NAME, STORE_INFO, key)).priceHistory;
-    return Plot.line(Object.entries(priceData), { stroke: "steelblue", tip: true }).plot({y: { grid: true }})
+    const priceData = (await sendToBackground({ name: "getItem", body: { key: key } })).item.priceHistory
+    return Plot.line(Object.entries(priceData).map((val: [string, number]) => [new Date(val[0]), val[1]]), { stroke: "steelblue", tip: true }).plot({y: { grid: true }, height: 500, width: 500})
 }
